move model def back to kserve
