+++ python/kserve/kserve/protocol/rest/timeseries/dataplane.py
@@ -18,13 +18,13 @@ from typing import Union, List
 
 from fastapi import Request, Response
 from starlette.datastructures import Headers
-from ...dataplane import DataPlane
-from huggingfaceserver.time_series_model import (
+from kserve.protocol.dataplane import DataPlane
+from .time_series_model import (
     HuggingFaceTimeSeriesModel,
     TimeSeriesModel,
 )
 
-from kserve.protocol.rest.timeseries.types import (
+from .types import (
     ForecastRequest,
     ForecastResponse,
     ErrorResponse,
