+++ python/kserve/kserve/protocol/rest/timeseries/time_series_model.py
